+++ axelor-production/src/main/java/com/axelor/apps/production/service/ManufOrderStockMoveService.java
@@ -259,6 +259,7 @@ public class ManufOrderStockMoveService {
         prodProduct.getProduct().getDescription(),
         qty,
         costPrice,
+        costPrice,
         prodProduct.getUnit(),
         stockMove,
         inOrOutType,